requirements.txt
streamlit
tensorflow==2.13.0
Pillow
opencv-python
matplotlib
numpy
openai
